refactor to ensure code doesn't depend on integration code
