+++ pkg/integration/README.md
@@ -37,15 +37,15 @@ If you find yourself doing something frequently in a test, consider making it a 
 
 There are three ways to invoke a test:
 
-1. go run pkg/integration/runner/main.go [<testname>...]
-2. go run pkg/integration/tui/main.go
+1. go run pkg/integration/cmd/runner/main.go [<testname>...]
+2. go run pkg/integration/cmd/tui/main.go
 3. go test pkg/integration/integration_test.go
 
 The first, the test runner, is for directly running a test from the command line. If you pass no arguments, it runs all tests.
 The second, the TUI, is for running tests from a terminal UI where it's easier to find a test and run it without having to copy it's name and paste it into the terminal. This is the easiest approach by far.
 The third, the go-test command, intended only for use in CI, to be run along with the other `go test` tests. This runs the tests in headless mode so there's no visual output.
 
-The name of a test is based on its path, so the name of the test at `pkg/integration/tests/commit/new_branch.go` is commit/new_branch. So to run it with our test runner you would run `go run pkg/integration/runner/main.go commit/new_branch`.
+The name of a test is based on its path, so the name of the test at `pkg/integration/tests/commit/new_branch.go` is commit/new_branch. So to run it with our test runner you would run `go run pkg/integration/cmd/runner/main.go commit/new_branch`.
 
 You can pass the KEY_PRESS_DELAY env var to the test runner in order to set a delay in milliseconds between keypresses, which helps for watching a test at a realistic speed to understand what it's doing. Or in the tui you can press 't' to run the test with a pre-set delay.
 
@@ -68,7 +68,7 @@ At the moment, all the deprecated test code lives in pkg/integration/deprecated.
 We should never write any new tests under the old method, and if a given test breaks because of new functionality, it's best to simply rewrite it under the new approach. If you want to run a test for the sake of watching what it does so that you can transcribe it into the new approach, you can run:
 
 ```
-go run pkg/integration/deprecated/tui/main.go
+go run pkg/integration/deprecated/cmd/tui/main.go
 ```
 
 The tests in the old format live in test/integration. In the old format, test definitions are co-located with the snapshots. The setup step is done in a `setup.sh` shell script and the `recording.json` file contains the recorded keypresses to be replayed during the test.
