+++ pkg/integration/components/test.go
@@ -81,7 +81,7 @@ func (self *IntegrationTest) SetupRepo(shell *Shell) {
 }
 
 // I want access to all contexts, the model, the ability to press a key, the ability to log,
-func (self *IntegrationTest) Run(gui integrationTypes.GuiAdapter) {
+func (self *IntegrationTest) Run(gui integrationTypes.GuiDriver) {
 	shell := NewShell()
 	assert := NewAssert(gui)
 	keys := gui.Keys()
